requirements.txt
streamlit
snowflake-ml-python
snowflake-snowpark-python
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
chromadb>=0.4.22
tiktoken>=0.5.2
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
pdfplumber>=0.10.0
PyPDF2>=3.0.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
pillow>=10.0.0
requests>=2.31.0
